[TSAR, Tool] Invoke LLVM IR processing action directly.
Clang Tooling support actions for sources only, so we process .ll files separately now.
